refactor: io managers and optional dependencies
