+++ docs/installation.md
@@ -13,3 +13,11 @@ uv add https://github.com/JasperHG90/dagster-pyiceberg/releases/download/v<VERSI
 ```
 
 You can find a list of versions / releases [here](https://github.com/JasperHG90/dagster-pyiceberg/releases).
+
+The following extras are available:
+
+- daft (for interoperability with Daft dataframes)
+- polars (for interoperability with Polars dataframes)
+- pandas (for interoperability with Pandas dataframes)
+
+Pyarrow is installed as a default dependency, so that IO Manager is always available.
