+++ src/dagster_pyiceberg/io_manager/__init__.py
@@ -1,0 +1,1 @@
+from dagster_pyiceberg.io_manager.base import IcebergIOManager as IcebergIOManager
