+++ tests/_db_io_manager/test_db_io_manager.py
@@ -27,7 +27,8 @@ from dagster import (
     materialize,
 )
 
-from dagster_pyiceberg import IcebergCatalogConfig, IcebergPyarrowIOManager
+from dagster_pyiceberg.config import IcebergCatalogConfig
+from dagster_pyiceberg.io_manager.arrow import IcebergPyarrowIOManager
 
 
 @pytest.fixture
