+++ tests/io_managers/test_arrow_io_manager.py
@@ -14,7 +14,8 @@ from dagster import (
 )
 from pyiceberg.catalog import Catalog
 
-from dagster_pyiceberg import IcebergCatalogConfig, IcebergPyarrowIOManager
+from dagster_pyiceberg.config import IcebergCatalogConfig
+from dagster_pyiceberg.io_manager.arrow import IcebergPyarrowIOManager
 
 
 @pytest.fixture
